:recycle: #187 - refactor: refactor destruction list pages structure
