--- frontend/src/components/index.ts
@@ -1,2 +1,0 @@
-/* eslint-disable prettier/prettier */
-export * from "./DestructionList/DestructionList";
