+++ frontend/src/pages/destructionlist/create/components/DestructionList/index.ts
@@ -1,0 +1,1 @@
+export * from "./DestructionList";
