+++ frontend/src/pages/destructionlist/detail/components/DestructionListAssignees/index.ts
@@ -1,0 +1,1 @@
+export * from "./DestructionListAssignees";
