+++ frontend/src/pages/destructionlist/detail/components/DestructionListEdit/index.ts
@@ -1,0 +1,1 @@
+export * from "./DestructionListEdit";
