+++ frontend/src/pages/destructionlist/detail/components/DestructionListProcessReview/index.ts
@@ -1,0 +1,1 @@
+export * from "./DestructionListProcessReview";
