+++ frontend/src/pages/destructionlist/detail/components/DestructionListProcessZaakReviewModal/index.ts
@@ -1,0 +1,1 @@
+export * from "./DestructionListProcessZaakReviewModal";
