+++ frontend/src/pages/destructionlist/detail/components/DestructionListToolbar/index.ts
@@ -1,0 +1,1 @@
+export * from "./DestructionListToolbar";
