+++ frontend/src/pages/destructionlist/detail/components/index.ts
@@ -1,0 +1,5 @@
+export * from "./DestructionListAssignees/DestructionListAssignees";
+export * from "./DestructionListEdit/DestructionListEdit";
+export * from "./DestructionListProcessReview/DestructionListProcessReview";
+export * from "./DestructionListProcessZaakReviewModal/DestructionListProcessZaakReviewModal";
+export * from "./DestructionListToolbar";
